src/api/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = "~=3.11.0"
readme = "README.md"
dependencies = [
    "fastapi>=0.110.1,<0.111",
    "uvicorn>=0.29.0,<0.30",
    "line-bot-sdk>=3.11.0,<4",
    "python-dotenv>=1.0.1,<2",
    "azure-cosmos>=4.6.0,<5",
    "pytz~=2025.2",
    "gunicorn>=22.0.0,<23",
    "azure-monitor-opentelemetry>=1.0.0,<2",
    "opentelemetry-instrumentation-fastapi>=0.40b0,<1",
    "langchain>=1.0,<2",
    "langchain-core>=1.0,<2",
    "langchain-openai>=1.0,<2",
    "langchain-mcp-adapters>=0.1.12,<0.2",
    "langgraph>=1.0,<2",
    "langgraph-checkpoint-postgres>=3.0.1,<4",
    "langsmith>=0.1.99",
    "google-auth>=2.40.3",
    "google-api-python-client>=2.171.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "psycopg[binary,pool]>=3.2.13",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.25.2,<0.26",
    "ruff>=0.11.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 127
